refactor: 결제/정산 금액 타입 BigDecimal 전환 (#57)
* refactor: 결제/정산 금액 타입 BigDecimal 전환

* refactor: BigDecimal 직렬화 안전성 및 원화 정수 입력 강제

+++ backend/src/main/java/com/back/team9/moyeota/domain/payment/client/TossConfirmResponse.java
@@ -1,9 +1,11 @@
 package com.back.team9.moyeota.domain.payment.client;
 
+import java.math.BigDecimal;
+
 public record TossConfirmResponse(
         String paymentKey,
         String orderId,
         String status,
-        Integer totalAmount
+        BigDecimal totalAmount
 ) {
 }
